delete feature websocket is pending

+++ lib/entity/Employee.dart
@@ -7,7 +7,7 @@ class Employee
   int? idTemp;
   int? id;
   String name;
-  int salary;
+  double salary;
   int? ts;
   int deleted;
